+++ Sources/KeyboardKit/Actions/KeyboardAction+Actions.swift
@@ -6,7 +6,7 @@
 //  Copyright © 2021 Daniel Saidi. All rights reserved.
 //
 
-import UIKit
+import Foundation
 
 /**
  This extension defines various actions that by default will