+++ Sources/KeyboardKit/Audio/StandardSystemAudioPlayer.swift
@@ -13,7 +13,21 @@ import AudioToolbox
  */
 public class StandardSystemAudioPlayer: SystemAudioPlayer {
     
-    public func playSystemAudio(_ id: UInt32) {
-        AudioServicesPlaySystemSound(id)
+    /**
+     Play a certain system audio sound.
+     **/
+    public func playSystemAudio(_ audio: SystemAudio) {
+        switch audio {
+        case .none: return
+        default: AudioServicesPlaySystemSound(audio.id)
+        }
     }
+}
+
+public extension StandardSystemAudioPlayer {
+    
+    /**
+     The standard player that is used for audio feedback.
+     */
+    static var shared: SystemAudioPlayer = StandardSystemAudioPlayer()
 }