+++ Sources/KeyboardKit/Audio/SystemAudio.swift
@@ -1,5 +1,5 @@
 //
-//  AudioFeedback.swift
+//  SystemAudio.swift
 //  KeyboardKit
 //
 //  Created by Daniel Saidi on 2019-10-15.
@@ -9,7 +9,7 @@
 import Foundation
 
 /**
- This enum contains various audio feedback types.
+ This enum contains various system audio types.
   
  You can call `trigger()` on either the type or an instance,
  to trigger the desired feedback.
@@ -20,7 +20,7 @@ import Foundation
  
  `TODO` Make this `Codable` (which requires Xcode 13) in 5.0.
 */
-public enum AudioFeedback: Equatable {
+public enum SystemAudio: Codable, Equatable, Identifiable {
     
     case
     input,
@@ -32,33 +32,27 @@ public enum AudioFeedback: Equatable {
     /**
      The system sound ID that corresponds to the feedback.
      */
-    public var systemId: UInt32? {
+    public var id: UInt32 {
         switch self {
         case .input: return 1104
         case .delete: return 1155
         case .system: return 1156
         case .custom(let value): return value
-        case .none: return nil
+        case .none: return 0
         }
     }
-    
-    /**
-     The standard player that is used for audio feedback.
-     */
-    public static var player: SystemAudioPlayer = StandardSystemAudioPlayer()
 }
 
 
 // MARK: - Public Functions
 
-public extension AudioFeedback {
+public extension SystemAudio {
     
     func trigger() {
-        AudioFeedback.trigger(self)
+        SystemAudio.trigger(self)
     }
     
-    static func trigger(_ feedback: AudioFeedback) {
-        guard let id = feedback.systemId else { return }
-        player.playSystemAudio(id)
+    static func trigger(_ audio: SystemAudio) {
+        StandardSystemAudioPlayer.shared.playSystemAudio(audio)
     }
 }